+++ src/tools/registerTools.ts
@@ -5,11 +5,13 @@ import { registerTerminalTools } from './terminalTools'
 import { registerSearchTool } from './searchTool'
 import { registerDiagnosticsTool } from './diagnosticsTool'
 import { registerGitContextTool } from './gitContextTool'
+import { registerTaskPlanTool } from '../agent/taskPlan'
 
 export function registerLanguageModelTools(context: vscode.ExtensionContext): void {
   registerFileTools(context)
   registerSearchTool(context)
   registerDiagnosticsTool(context)
   registerGitContextTool(context)
+  registerTaskPlanTool(context)
   registerTerminalTools(context)
 }